requirements.txt
supabase==2.10.0
streamlit==1.41.1
gotrue==2.11.0
resend==2.5.1
openai==1.54.3
google-generativeai==0.7.2 
langchain-community==0.3.5
langchain==0.3.7
langchain-experimental==0.3.3